+++ js/includeLocal.js
@@ -1,8 +1,9 @@
 function includeLocalCSS(filename) {
-    // $("head").append(`<link href="css/vendor/${filename}" rel="stylesheet">`);
-    document.write(`<link href='css/vendor/${filename}' rel="stylesheet">`);
+  document.write(`<link href='css/vendor/${filename}' rel="stylesheet">`);
 }
 
 function includeLocalJS(name) {
-    document.write(`<script type='text/javascript' src='js/vendor/${name}'><\/script>`);
+  document.write(
+    `<script type='text/javascript' src='js/vendor/${name}'><\/script>`
+  );
 }