+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog][Keep a Changelog url], and this projec
 1. Add CSS3 to Built With section of README.
 1. Add JavaScript to Built With section of README.
 1. Add jQuery to Built With section of README.
+1. Remove commented code from JS script.
 
 ## [1.0.3] (30 May 2023)